% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrainedLS_osqp.R
\name{constrainedLS_osqp}
\alias{constrainedLS_osqp}
\title{OSQP quadratic programming solver to solve a constrained least squares problem}
\usage{
constrainedLS_osqp(
  X,
  y,
  nonnegative = rep(TRUE, ncol(X)),
  x.start = NULL,
  y.start = NULL
)
}
\arguments{
\item{X}{the covariate (design) matrix}

\item{y}{vector of values of the dependent variable}

\item{nonnegative}{a vector of TRUE/FALSE representing whether
to apply nonnegative constraints on each of the model's coefficients}

\item{x.start}{the starting value for x}

\item{y.start}{the starting value for the expected value of y}
}
\value{
object retuned by OSQP Solve()
}
\description{
This function utilizes the OSQP package to carry out a constrained least squares regression.
}
