% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.l0adri.R
\name{cv.l0adri}
\alias{cv.l0adri}
\title{Cross Validation to tune the regularization parameter lambda for L0 adaptive ridge models}
\usage{
cv.l0adri(
  X,
  y,
  weights = rep(1, nrow(X)),
  family = c("gaussian(identity)", "poisson(identity)", "Gamma(identity)",
    "inv.gaussian(identity)", "binomial(logit)"),
  lambdas,
  nfolds = 10,
  maxit = 100,
  seed = 1,
  nonnegative = rep(T, ncol(X)),
  algo = c("osqp", "clipping", "scaling"),
  tol = 1e-08,
  thresh = 0.001,
  Plot = TRUE
)
}
\arguments{
\item{X}{the covariate matrix}

\item{y}{vector of values of the dependent variable}

\item{weights}{observation weights}

\item{family}{desired noise model and link function; all default R families
are supported, e.g. gaussian(identity), poisson(identity), Gamma(identity),
inv.gaussian(identity) or binomial(logit)}

\item{lambdas}{vector with values of the regularization parameter lambda for
which to perform cross-validation}

\item{nfolds}{the number of folds to use during cross validation}

\item{maxit}{the maximum number of times to iterate the adaptive ridge fits}

\item{seed}{random seed}

\item{nonnegative}{a vector of TRUE/FALSE representing whether
to apply nonnegative constraints for each of the coefficients}

\item{algo}{the algorithm used to apply the constraints. Can be 
osqp (quadratic programming), clipping or scaling}

\item{tol}{convergence tolerance}

\item{thresh}{coefficients with absolute value below this threshold will be
set to zero}

\item{Plot}{whether or not to plot the cross validation error in function of the lambda values}
}
\value{
list with elements lam.min (the optimal value of lambda that gives
  the minimal weighted mean square error), cv.error (weighted mean square error for different lambda values) and error
  (weighted total sums of squares for different folds and lambda values)
}
\description{
This function uses k-fold cross validation for l0adri and returns the value
of the tuning parameter lambda that yields the lowest mean square prediction
error.
}
\examples{
lam1=cv.l0adri(X=X,y=y,lam=l,family=gaussian(identity),nonnegative = rep(TRUE,ncol(X))


}
