% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l0adri.R
\name{l0adri}
\alias{l0adri}
\title{L0-penalized GLM fit using an iterative adaptive ridge regression approach}
\usage{
l0adri(
  X,
  y,
  weights = rep(1, nrow(X)),
  family = c("gaussian(identity)", "poisson(identity)", "Gamma(identity)",
    "inv.gaussian(identity)", "binomial(logit)"),
  lam = 0,
  nonnegative = rep(T, ncol(X)),
  algo = c("osqp", "clipping", "scaling"),
  miniter = 5,
  maxit = 30,
  tol = 1e-08,
  thresh = 0.001
)
}
\arguments{
\item{X}{the covariate (design) matrix (can be dense or sparse, i.e. of class dgCMatrix)}

\item{y}{vector of values of the dependent variable}

\item{weights}{observation weights}

\item{family}{desired noise model and link function; all default R families
are supported, e.g. gaussian(identity), poisson(identity), Gamma(identity),
inv.gaussian(identity) or binomial(logit)}

\item{lam}{regularization parameter lambda, which will be multiplied with the
adaptive penalty weights during each iteration of the L0 adaptive ridge algorithm
to give a per-coefficient adaptive ridge penalty}

\item{nonnegative}{a vector of TRUE/FALSE representing whether
to apply nonnegative constraints on each of the model's coefficients}

\item{algo}{the algorithm used to apply the constraints. Can be 
osqp (quadratic programming), clipping or scaling}

\item{miniter}{the minimum number of iterations of the iterative ridge after
which the observation weights of the IRLS GLM algorithm will be updated}

\item{maxit}{the maximum number of iterations of the iterative adaptive ridge algorithm}

\item{tol}{convergence tolerance}

\item{thresh}{coefficients with absolute value below this threshold will be
set to zero}
}
\value{
list with elements beta (L0 adaptive ridge coefficients) and weights (the optimized weights of the IRLS GLM fitting algorithm)
}
\description{
This function fits an L0-penalized GLM models using an iterative adaptive ridge
regression approach under the specified noise model & link function, with or
without nonnegativity or box constraints (upper & lower bounds) on the fitted
coefficients.
}
\examples{
fit <- l0adri(X = sim$X, y = sim$y, family = poisson(identity),
                   algo = "osqp", lam = 0,
                   nonnegative = rep(TRUE,nrow(X)))
l0araR_wgaus_fit <- l0adri(X = sim$X, y = sim$y, weights = 1/(sim$y+1),
                           family = gaussian(identity), lam = 1, 
                           nonnegative = TRUE, maxit = 25, tol = 1E-8)


}
