\name{L0adri1.0-package}
\alias{L0adri1.0-package}
\alias{L0adri1.0}
\docType{package}
\title{
\packageTitle{L0adri1.0}
}
\description{
\packageDescription{L0adri1.0}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{L0adri1.0}
\packageIndices{L0adri1.0}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{L0adri1.0}

Maintainer: \packageMaintainer{L0adri1.0}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
