% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_adridge.R
\name{lm_adridge}
\alias{lm_adridge}
\title{Function to fit a single adaptive ridge regression with or without nonnegativity or box constraints}
\usage{
lm_adridge(
  X,
  y,
  weights,
  lambda = 1,
  beta = rep(1, ncol(X)),
  thresh = 1e-04,
  nonnegative = rep(T, ncol(X)),
  algo = c("osqp", "clipping", "scaling")
)
}
\arguments{
\item{X}{the covariate matrix}

\item{y}{vector of values of the dependent variable}

\item{weights}{observation weights of the IRLS GLM fit in each iteration}

\item{beta}{coefficient values from which the adaptive ridge penalty weights=1/beta^2 are calculated; each coefficients will receive an L2-penalty of lam * the adaptive penalty weight.}

\item{thresh}{coefficients with absolute value below this threshold will be
set to zero}

\item{nonnegative}{a vector of TRUE/FALSE representing whether
to apply nonnegative constraints on each of the coefficients}

\item{algo}{the algorithm used to apply the constraints. Can be 
osqp (quadratic programming), clipping or scaling}

\item{lam}{the value of tuning parameter lambda for the ridge
regression}
}
\value{
vector with coefficients
}
\description{
This function fits a least squares adaptive ridge regression model using squared L2 norm regularization on each coefficient of lam / beta^2, where value lam and the initial coefficient estimates beta can both be specified.
An efficient linear solver is used that adapts to the different arguments given, e.g. whether or not covariate matrix X is sparse, whether or not nonnegativity or box constraints are required,
and the size of the covariate matrix. Three different methods to deal with nonnegativity or box constraints are provided (osqp=quadratic programming, clipping or scaling).
When covariate matrix X is sparse efficient sparse solvers from the Eigen package are used.
}
